Introducing auto-generated configs from class/function signatures
