+++ pyproject.toml
@@ -22,6 +22,7 @@ torch-optimizer = "^0.3.0"
 rich = "12.6.0"
 timm = "0.6.11"
 hydra-core = "^1.3.2"
+hydra-zen = "^0.11.0"
 
 [tool.poetry.group.dev.dependencies]
 flake8 = "^6.1.0"
